% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{AddNum}
\alias{AddNum}
\title{AddNum}
\usage{
AddNum(x, y)
}
\arguments{
\item{x}{First Num}

\item{y}{Second Num}
}
\value{
The sum of the numbers x and y
}
\description{
Adds Num
}
\examples{
AddNum(1, 2)
}

