% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DivNum}
\alias{DivNum}
\title{DivNum}
\usage{
DivNum(x, y)
}
\arguments{
\item{x}{Numerator}

\item{y}{Denominator}
}
\value{
Division result
}
\description{
Div num
}
\examples{
DivNum(1, 2)
}

