% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_names.R
\name{list_package_args}
\alias{list_package_args}
\title{creates list of log files}
\usage{
list_package_args(path)
}
\arguments{
\item{path}{to the package test file}
}
\description{
creates list of log files
}
