\name{stopgo}
\alias{stopgo}
\title{Stop or Go?}
\description{a simple function that based on the color sampled decides if you should stop or go. Similar to a die roll, it relies on chance to select the movement. Due to the sample space only being 2, the number input must be 1.}
\usage{stopgo(x)}
\details{input must be 1!
\value{returned value with either be STOP! or GO!}
\references{None at the time}
\author{Aira Contreras}
\note{please, just use 1}
\examples{stopgo(1)}

