% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveylevels.R
\name{surveylevels}
\alias{surveylevels}
\title{Converts Survey Character Levels to Numeric}
\usage{
surveylevels(data, column, ordered_levels)
}
\arguments{
\item{data}{A data frame containing the column to be converted.}

\item{column}{A string specifying the name of the column to convert. The column should contain
character values that match the provided ordered levels.}

\item{ordered_levels}{A character vector containing the desired order of levels. The first level in the
vector will be assigned the value 1, the second level will be assigned the value 2, and so on.}
}
\value{
The modified data frame with the specified column replaced by its numeric equivalents.
}
\description{
This function converts specified character levels in a given column of a data frame to numeric values.
The conversion is based on provided order of levels, which ensures that the numeric values
correspond to the intended ranking or hierarchy. The function replaces the original column with the
numeric equivalents while validating that all levels in the column are part of the provided order.
}
\examples{
data <- data.frame(proficiency = c("Well", "Not at all", NA, "Not well", "Very well"))
response_scale <- c("Not at all", "Not well", "Well", "Very well")
data <- surveylevels(data, column = "proficiency", ordered_levels = response_scale)
print(data)
}
