\name{testproject-package}
\alias{testproject-package}
\alias{testproject}
\docType{package}
\title{\packageTitle{testproject}}
\description{\packageDescription{testproject}}
\section{Package Content}{\packageIndices{testproject}}
\author{\packageAuthor{testproject}}
\section{Maintainer}{\packageMaintainer{testproject}}
\keyword{package}
